Labs: Deassociate lab numbers from content
Labs are commonly rotated and replaced each year and updating them
requires creating commits that remove files or large amounts of content
from files. This commit moves the contents of the labs to another
directory, called "content", and maintains the appearance of the old
structure using symbolic links.

The approach of using symbolic links makes rotations much easier, as
only the symbolic link can be changed to point to another location.

Signed-off-by: Darius Mihai <dariusmihaim@gmail.com>