Add solutions and modify `README.md` for lab01
Add for the exercises to the solution/ directory.

The `README.md` is hard to read at times, mainly because of crowded
paragraphs and values such as `0x564c364482a0`, positioned inline
with other text, and without being emphasized with formatting.
Update the formatting of the file to allow for better visibility.

Add a small example at the beginning of the `README.md` for value
prefixes in C, such as `0x` or `0b`.

Add BSD-3 License identifier to solution code to avoid
overly aggressive checkpatch errors.

Signed-off-by: RCX777 <robertcnst02@gmail.com>