Update PR based on reviewer comments
As @teodutu suggested, the `utils/` directory was unnecessary, as
there is a common directory which stores the build files. Removsd
the `utils/` directory in the lab and updates the Makefiles in the
`solution/` directory, so that they use the files inside the
common location.

Modifies the formatting of `README.md`, as there is no need
for **bold** in titles, as they already are rendered in a
larger font.

Removes the unneeded "Solutions" chapter from `README.md`, as
it is obvious where the solutions are located.

Signed-off-by: RCX777 <robertcnst02@gmail.com>